added subfolders for components
