--- client/components/App.test.js
@@ -1,10 +1,0 @@
-// import React from 'react'
-// import { Provider } from 'react-redux'
-// import { screen, render } from '@testing-library/react'
-
-// import App from './App'
-// import store from '../store'
-
-// jest.mock('../actions')
-
-test.todo('Add tests for App component')
