+++ client/components/Auth/Authenticated.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+
+import { isAuthenticated } from '../../auth-utils'
+
+export function IfAuthenticated ({ children }) {
+  return isAuthenticated()
+    ? <>{children}</>
+    : null
+}
+
+export function IfNotAuthenticated ({ children }) {
+  return !isAuthenticated()
+    ? <>{children}</>
+    : null
+}
