+++ client/components/Friends/Friends.jsx
@@ -1,10 +1,12 @@
 import React, { useEffect } from 'react'
 import { Link } from 'react-router-dom'
-import { fetchFriends } from './appHelper'
-import { setCurrentFriend } from '../actions'
 import { connect } from 'react-redux'
-import Icon from './Icon'
-import DataRowText from './DataRowText'
+
+import { fetchFriends } from './friendsHelper'
+import { setCurrentFriend } from '../../actions'
+
+import Icon from '../Icon/Icon'
+import DataRowText from '../DataRow/DataRowText'
 
 function Friends (props) {
   const id = 10001
@@ -27,16 +29,11 @@ function Friends (props) {
       <p>My existing friends:</p>
       {!props.friends.length && (
         <p>Instructions for adding a new friend</p>)}
-      {/* <form>
-        <input type="text" />
-      </form> */}
       {props.friends.map(friend => (
         <div className='data-row' key={friend.id} onClick={() => props.dispatch(setCurrentFriend(friend))}>
           <Link to={`/friends/${friend.id}`}><Icon style='user' /></Link>
           <DataRowText link={`/friends/${friend.id}`} title={friend.nickname} subtitle={`${friend.name} ${friend.lastName}`} />
         </div>
-        // <li key={friend}><Link to={`/friends/${friend.id}`}><button>icon</button>{friend.nickname}
-        //   <br /> {friend.name} {friend.lastName} </Link>X</li>
       ))}
     </>
   )
