+++ client/components/Icon/Icon.test.js
