+++ client/components/Logo/Logo.test.js
