+++ client/components/Songs/Songs.jsx
@@ -1,12 +1,13 @@
 import React, { useEffect } from 'react'
 import { useHistory } from 'react-router'
-import { connect } from 'react-redux'
 import { Link } from 'react-router-dom'
-import Icon from './Icon'
-import DataRowText from './DataRowText'
-import Button from './Button'
-// import getUserSongs from '../actions'
-import { setSongs } from '../actions'
+import { connect } from 'react-redux'
+
+import { setSongs } from '../../actions'
+
+import Icon from '../Icon/Icon'
+import DataRowText from '../DataRow/DataRowText'
+import Button from '../Button/Button'
 
 function Songs (props) {
   const history = useHistory()
@@ -14,12 +15,6 @@ function Songs (props) {
   // const { nickname } = props.user
   const nickname = 'superman'
   const { songs } = props
-  // const songs = [
-  //   { id: 1, user_id: 10001, genre: 'Indie Rock', mediaName: 'Hengelo', artist: 'Spring Offensive', link: 'https://open.spotify.com/track/4rqpg85XNApASjAvqjHlb1?si=2bdc00343f3e47f2' },
-  //   { id: 2, user_id: 10001, genre: 'Indie Rock', mediaName: 'By Design', artist: 'Tigercub', link: 'https://open.spotify.com/track/6ICdz2wvVMDC4u801OwHA2?si=8e214d39012c4685' },
-  //   { id: 3, user_id: 10001, genre: 'Indie Rock', mediaName: 'High Strings', artist: 'Noise Punk', link: 'https://open.spotify.com/track/38HaKBYwrqnhrjf9sqJbfc?si=e9bc96d2247c4347' },
-  //   { id: 4, user_id: 10001, genre: 'Funky House', mediaName: 'I look to you (feat. Kimbra)', artist: 'Miami Horror', link: '' }
-  // ]
 
   useEffect(() => {
     props.dispatch(setSongs(userId))
